pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "matplotlib>=3.10.1",
    "nltk>=3.9.1",
    "numpy>=2.2.5",
    "openai>=1.76.0",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "scikit-learn>=1.6.1",
    "spacy>=3.8.5",
    "streamlit>=1.44.1",
    "wordcloud>=1.9.4",
]
